feat: try to support darwin
