+++ libcore/log_linux.go
@@ -1,10 +1,6 @@
-//go:build unix
-
 package libcore
 
-import (
-	"golang.org/x/sys/unix"
-)
+import "golang.org/x/sys/unix"
 
 func dup(oldFd, newFd, flags int) error {
 	return unix.Dup3(oldFd, newFd, flags)
