+++ libcore/log_stub.go
@@ -1,4 +1,4 @@
-//go:build !unix
+//go:build !(linux || darwin)
 
 package libcore
