+++ libcore/tun_name_darwin.go
@@ -1,0 +1,13 @@
+package libcore
+
+import (
+	"golang.org/x/sys/unix"
+)
+
+func tunnelName(fd int32) (string, error) {
+	return unix.GetsockoptString(
+		int(fd),
+		2, /* #define SYSPROTO_CONTROL 2 */
+		2, /* #define UTUN_OPT_IFNAME 2 */
+	)
+}
