+++ libcore/tun_name_stub.go
@@ -1,0 +1,11 @@
+//go:build !(linux || darwin)
+
+package libcore
+
+import (
+	"os"
+)
+
+func tunnelName(fd int32) (string, error) {
+	return "", os.ErrInvalid
+}
